druppie/mcp-servers/module-azure-datalake/requirements.txt
fastmcp>=0.1.0
httpx>=0.25.0
uvicorn>=0.24.0
pyyaml>=6.0.0
azure-storage-file-datalake>=12.14.0
azure-identity>=1.15.0
pandas>=2.0.0
pyarrow>=14.0.0
